# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: RALD ELIMU — African Education Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes/{classId}/timetable:
    get:
      operationId: getClassTimetable
      tags: [timetables]
      summary: Get timetable for a class
      parameters:
        - $ref: "#/components/parameters/classId"
      responses:
        "200":
          description: Class timetable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableResponse"
    post:
      operationId: addTimetableEntry
      tags: [timetables]
      summary: Add a timetable entry
      parameters:
        - $ref: "#/components/parameters/classId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTimetableEntryBody"
      responses:
        "201":
          description: Created timetable entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
  /timetable/{entryId}:
    delete:
      operationId: deleteTimetableEntry
      tags: [timetables]
      summary: Delete a timetable entry
      parameters:
        - in: path
          name: entryId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Attendance ─────────────────────────────────────────────────────────────
...
  /verify/{hash}:
    get:
      operationId: verifyCredential
      tags: [verification]
      summary: Publicly verify an academic credential
      parameters:
        - in: path
          name: hash
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /credentials:
    post:
      operationId: issueCredential
      tags: [verification]
      summary: Issue a digital credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCredentialBody"
      responses:
        "201":
          description: Issued credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
  /students/{studentId}/credentials:
    get:
      operationId: getStudentCredentials
      tags: [verification]
      summary: Get all credentials for a student
      parameters:
        - $ref: "#/components/parameters/studentId"
      responses:
        "200":
          description: Student credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialListResponse"

  # ── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  parameters:
...
    studentId:
      in: path
      name: studentId
      required: true
      schema:
        type: string
...
    classId:
      in: path
      name: classId
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    # ── Enums ───────────────────────────────────────────────────────────────
...
    CredentialType:
      type: string
      enum: [result, certificate, diploma, degree, transcript, award]
    DayOfWeek:
      type: string
      enum: [monday, tuesday, wednesday, thursday, friday, saturday]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    TimetableEntry:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        subjectId:
          type: string
        subjectName:
          type: string
          nullable: true
        teacherId:
          type: string
          nullable: true
        teacherName:
          type: string
          nullable: true
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
          nullable: true
      required: [id, classId, subjectId, dayOfWeek, startTime, endTime]
    TimetableResponse:
      type: object
      properties:
        classId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TimetableEntry"
      required: [classId, entries]
    AddTimetableEntryBody:
      type: object
      properties:
        subjectId:
          type: string
        teacherId:
          type: string
        dayOfWeek:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
      required: [subjectId, dayOfWeek, startTime, endTime]

    # ── Attendance ──────────────────────────────────────────────────────────
...
    Credential:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: string
          nullable: true
        schoolId:
          type: string
        schoolName:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/CredentialType"
        title:
          type: string
        issuedAt:
          type: string
          format: date-time
        verificationHash:
          type: string
        verified:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, schoolId, type, title, issuedAt, verificationHash, verified, createdAt]
    CredentialListResponse:
      type: object
      properties:
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/Credential"
        total:
          type: integer
      required: [credentials, total]
    IssueCredentialBody:
      type: object
      properties:
        studentId:
          type: string
        type:
          $ref: "#/components/schemas/CredentialType"
        title:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required: [studentId, type, title]
    VerificationResult:
      type: object
      properties:
        verified:
          type: boolean
        credential:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/Credential"
            - type: "null"
        message:
          type: string
      required: [verified, credential, message]

    # ── Dashboard ────────────────────────────────────────────────────────────
...